Bug Fix Criar Queimadas (Still Not 100%)

+++ app/src/main/java/com/example/chamasegura/retrofit/tabels/Queimadas.kt
@@ -1,7 +1,7 @@
 package com.example.chamasegura.retrofit.tabels
 
 data class Queimadas(
-    val location: Int,
+    val location: Long,
     val type: String,
     val date: String,
     val reason: String
